% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc-data.R
\docType{data}
\name{adae}
\alias{adae}
\title{Adverse Event Analysis Dataset example dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 92 columns.
}
\usage{
adae
}
\description{
\link{adae} is a re-export of the \link[random.cdisc.data:cadae]{random.cdisc.data::cadae} dataset, included in
\code{{dtlg}} for function usage illustration and testing.
}
\examples{
adae

}
\keyword{datasets}
