% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc-data.R
\docType{data}
\name{adlb}
\alias{adlb}
\title{ADaM Basic Data Structure (BDS) example dataset}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 8400 rows and 102 columns.
}
\usage{
adlb
}
\description{
\link{adlb} is a re-export of the \link[random.cdisc.data:cadlb]{random.cdisc.data::cadlb} dataset, included in
\code{{dtlg}} for function usage illustration and testing.
}
\examples{
adlb

}
\keyword{datasets}
