% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc-data.R
\docType{data}
\name{adsl}
\alias{adsl}
\title{Subject-Level Analysis Dataset (ADSL) example dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 55 columns.
}
\usage{
adsl
}
\description{
\link{adsl} is a re-export of the \link[random.cdisc.data:cadsl]{random.cdisc.data::cadsl} dataset, included in
\code{{dtlg}} for function usage illustration and testing.
}
\examples{
adsl

}
\keyword{datasets}
