% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtlg-package.R
\docType{package}
\name{dtlg-package}
\alias{dtlg}
\alias{dtlg-package}
\title{dtlg: A Performance-Focused Package for Clinical Trial Tables}
\description{
Create pharmaceutical tables by using data.table as backend.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://AscentSoftware.github.io/dtlg/}
}

}
\author{
\strong{Maintainer}: Ramiro Magno \email{ramiro.morgado@ascent.io}

Authors:
\itemize{
  \item Max Ebenezer-Brown \email{max.ebenezer-brown@ascent.io}
  \item Max Norman \email{max.norman@ascent.io}
  \item Xinye Li \email{xinye.li@ascent.io}
  \item Anja Peebles-Brown \email{anna.peebles-brown@ascent.io}
}

}
\keyword{internal}
