% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_label.R
\name{label}
\alias{label}
\title{Retrieve the label of an object}
\usage{
label(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
The label attribute (string) associated with object passed in \code{x} or
\code{NULL} if the label attribute does not exist.
}
\description{
\code{\link[=label]{label()}} gets the attached label to an object.
}
\examples{
label(1)
label(with_label(1, "my label"))

}
\seealso{
\code{\link[=with_label]{with_label()}}
}
