% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indentation.R
\name{nbsp}
\alias{nbsp}
\title{Generate Non-Breaking Spaces for HTML Output}
\usage{
nbsp(n = 1L)
}
\arguments{
\item{n}{Number of non-breaking spaces. Defaults to \code{1}.}
}
\value{
A character string containing \code{n} HTML non-breaking spaces
(\verb{&nbsp;}).
}
\description{
\code{\link[=nbsp]{nbsp()}} generates a string of HTML non-breaking spaces (\verb{&nbsp;}).
}
\examples{
# One non-breaking space
nbsp()
nbsp(1)

# Several non-breaking spaces
nbsp(n = 2)
nbsp(n = 5)

# When `n` is zero, an empty string is returned.
nbsp(n = 0)

}
\keyword{internal}
