% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with_label.R
\name{with_label}
\alias{with_label}
\title{Add a label attribute to an object}
\usage{
with_label(x, label)
}
\arguments{
\item{x}{An R object.}

\item{label}{A label provided as a single string.}
}
\value{
\code{x} labeled by \code{label}.
}
\description{
Add a label attribute to an object
}
\examples{
label(1)
label(with_label(1, "my label"))

}
\seealso{
\code{\link[=label]{label()}}
}
